Added support for specialized select parameters.
This can be used to implement special ordering of the search,
such as putting the results with prefix match prior to results
based on general ilike '%pattern%' match.
Example code from a controller that gives priority to exact
prefix matches for the name field, in addition to the full
text search:
such as putting the results with prefix match prior to results
based on general ilike '%pattern%' match.
Example code from a controller that gives priority to exact
prefix matches for the name field, in addition to the full
text search:

 record_select(:search_on => [:name],
   :full_text_search => true,
   :order_by => "pos nulls last, name")

 def record_select_select
   "*,case position(lower(#{ActiveRecord::Base.connection.quote(params[:search])}) in lower(name)) when 1 then true else null end as pos"
 end
 record_select(:search_on => [:name],
   :full_text_search => true,
   :order_by => "pos nulls last, name")

 def record_select_select
   "*,case position(lower(#{ActiveRecord::Base.connection.quote(params[:search])}) in lower(name)) when 1 then true else null end as pos"
 end